+++ clients/src/main/java/org/apache/kafka/common/network/BrokerEndPoint.java
@@ -17,9 +17,6 @@
 package org.apache.kafka.common.network;
 
 import java.util.Objects;
-import java.util.Optional;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 /**
  * BrokerEndPoint is used to connect to specific host:port pair.
@@ -30,8 +27,6 @@ import java.util.regex.Pattern;
  */
 public class BrokerEndPoint {
 
-    private static final Pattern URI_PARSE_EXP = Pattern.compile("\\[?([0-9a-zA-Z\\-%._:]*)]?:([0-9]+)");
-
     private final int id;
     private final String host;
     private final int port;
@@ -40,14 +35,6 @@ public class BrokerEndPoint {
         this.id = id;
         this.host = host;
         this.port = port;
-    }
-
-    /**
-     * This constructor is only used by the static parseHostPort method, which acts as a factory method
-     *  to create a BrokerEndPoint from a host:port string.
-     */
-    private BrokerEndPoint(String host, int port) {
-        this(0, host, port);
     }
 
     public int id() {
@@ -60,22 +47,6 @@ public class BrokerEndPoint {
 
     public int port() {
         return port;
-    }
-
-    /**
-     * BrokerEndPoint URI is host:port or [ipv6_host]:port
-     * Note that unlike EndPoint (or listener) this URI has no security information.
-     */
-    public static Optional<BrokerEndPoint> parseHostPort(String connectionString) {
-        Matcher matcher = URI_PARSE_EXP.matcher(connectionString);
-        if (matcher.matches()) {
-            try {
-                return Optional.of(new BrokerEndPoint(matcher.group(1), Integer.parseInt(matcher.group(2))));
-            } catch (NumberFormatException e) {
-                // Ignore
-            }
-        }
-        return Optional.empty();
     }
 
     @Override
